+++ gitbook2pdf.net/Program.cs
@@ -1,16 +1,17 @@
-﻿namespace gitbook2pdf.net;
+﻿using Microsoft.Extensions.DependencyInjection;
+using SimpleHtmlToPdf.Interfaces;
+
+namespace gitbook2pdf.net;
 
 class Program
 {
     static async Task Main(string[] args)
     {
         // License warning
-        Console.WriteLine("");
-        Console.WriteLine("");
         Console.WriteLine("gitbook2pdf.NET Copyright (C) 2025 alciarro93");
-        Console.WriteLine("This program comes with ABSOLUTELY NO WARRANTY; for details type `show w'.");
+        Console.WriteLine("This program comes with ABSOLUTELY NO WARRANTY.");
         Console.WriteLine("This is free software, and you are welcome to redistribute it");
-        Console.WriteLine("under certain conditions; type `show c' for details.");
+        Console.WriteLine("under certain conditions.");
         Console.WriteLine("");
         Console.WriteLine("");
 
@@ -31,12 +32,23 @@ class Program
 
         if (!string.IsNullOrEmpty(gitbook_url))
         {
+            //Composition root
+            IServiceProvider services = ConfigureServices();
+            var htmlPDFServ = services.GetRequiredService<IConverter>();
+            
             // start convertion to PDF
-            await GitBookUtils.GitBookToPdf(gitbook_url);
+            await GitBookUtils.GitBookToPdf(gitbook_url, htmlPDFServ);
         }
         
         // stop console
         Console.WriteLine("Press any key to close");
         Console.Read();
+    }
+
+    private static IServiceProvider ConfigureServices() {
+        IServiceCollection services = new ServiceCollection();
+        services.AddSimpleHtmlToPdf();
+        
+        return services.BuildServiceProvider();
     }
 }