+++ CHANGELOG.md
@@ -1,17 +1,22 @@
 # Changelog
 
-## v1.0.0
+## v1.0.1
+### Added or Changed
+- Added new HTML to PDF package (SimpleHtmlToPdf)
+- Added "assets/css/style.css" to style the output pdf
+### Removed
+- All the references of the old package for HTML to PDF convertion (HiQPdf.NetCore free trial)
 
+
+## v1.0.0
 ### Added or Changed
 - Added this changelog :)
 - Added LICENSE.txt
 - Added checkboxes for major features on roadmap
-
 ### Removed
 - TODO comments
 
 
 ## v0.0.9
-
 ### Added or Changed
 - First public commit